requirements.txt
fastapi
uvicorn
llama-cpp-python
sentence-transformers==2.2.2
--extra-index-url https://download.pytorch.org/whl/cpu
numpy
pydantic
huggingface_hub==0.14.1
pymongo[srv]
qdrant-client
